+++ internal/domain/event/repository.go
@@ -2,9 +2,9 @@ package event
 
 type Repository interface {
 	FindAll() (*[]Event, error)
-	FindOne(int) (*Event, error)
+	FindOne(uint64) (*Event, error)
 	FindByCoords(float64, float64, float64) (*[]Event, error)
 	PostOne(*Event) error
-	UpdateOne(int, *Event) error
-	DeleteOne(int) error
+	UpdateOne(uint64, *Event) error
+	DeleteOne(uint64) error
 }
